Move under "oneoff" package and ktlint format
